investment-team: hoist orchestrator dataclasses + pure helpers to sibling module
Fix #3 of the latest review. ``orchestrator.py`` was carrying twelve
top-level concepts that weren't ``StrategyLabOrchestrator``: three
outcome dataclasses (``_MarketDataFetch``, ``_AlignmentLoopOutcome``,
``_VerificationOutcome``), six pure helpers previously hoisted off the
class (``_merge_risk_limits_tighten_only``, ``_daily_returns_from_trades``,
``_equity_to_returns``, ``_closes_to_equity``, ``_parse_bar_date``,
``_resolve_vix_provider``), and three module-level helpers
(``_maybe_attach_coverage_report``, ``_format_execution_diagnostics``,
``_apply_veto_to_acceptance_reason``).

All twelve move into ``_orchestrator_helpers.py``. ``orchestrator.py``
re-exports them at the bottom of the file so existing imports like
``from investment_team.strategy_lab.orchestrator import _MarketDataFetch``
continue to work — no changes needed in ``zero_trade_repair.py``, tests,
or external docs that reference these names.

One test (``test_maybe_attach_coverage_report_runs_stage_when_gate_on``)
monkeypatched ``run_coverage_stage`` on the orchestrator module. The
helper now lives in ``_orchestrator_helpers``, so the lookup site
moved with it; the test now patches the correct module.

The orchestrator file is now coordination-focused: the class definition
+ phase methods, plus a trailing re-export block. A new reader opens
it expecting to read the coordinator and meets the coordinator first.
